requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
email-validator==2.1.1
sqlalchemy==2.0.30
Werkzeug==2.3.8
Jinja2==3.1.3

